src/pgsqlmgr/db.py
"""PostgreSQL database operations and management."""

import psycopg2
from pathlib import Path
from typing import Optional, List, Dict, Any
from rich.console import Console

from .config import HostConfig, LocalHost, SSHHost

console = Console()


class DatabaseManager:
    """Manage PostgreSQL database operations."""
    
    def __init__(self, host_config: HostConfig):
        """Initialize database manager with host configuration."""
        self.config = host_config
        self._connection: Optional[psycopg2.extensions.connection] = None
    
    def connect(self) -> psycopg2.extensions.connection:
        """
        Establish connection to PostgreSQL database.
        
        Returns:
            psycopg2 connection object
            
        Raises:
            psycopg2.Error: If unable to connect to database
        """
        # Implementation placeholder
        console.print("[yellow]⚠️  Database connection not implemented yet[/yellow]")
        raise NotImplementedError("Database connection functionality coming soon")
    
    def test_connection(self) -> bool:
        """
        Test database connection.
        
        Returns:
            True if connection successful, False otherwise
        """
        # Implementation placeholder
        console.print("[yellow]⚠️  Connection test not implemented yet[/yellow]")
        return False
    
    def list_databases(self) -> List[str]:
        """
        List all databases on the PostgreSQL instance.
        
        Returns:
            List of database names
        """
        # Implementation placeholder
        console.print("[yellow]⚠️  Database listing not implemented yet[/yellow]")
        return []
    
    def database_exists(self, database_name: str) -> bool:
        """
        Check if a database exists.
        
        Args:
            database_name: Name of the database to check
            
        Returns:
            True if database exists, False otherwise
        """
        # Implementation placeholder
        console.print(f"[yellow]⚠️  Would check if database '{database_name}' exists[/yellow]")
        return False
    
    def create_database(self, database_name: str) -> None:
        """
        Create a new database.
        
        Args:
            database_name: Name of the database to create
            
        Raises:
            psycopg2.Error: If database creation fails
        """
        # Implementation placeholder
        console.print(f"[yellow]⚠️  Would create database '{database_name}'[/yellow]")
        raise NotImplementedError("Database creation functionality coming soon")
    
    def drop_database(self, database_name: str, force: bool = False) -> None:
        """
        Drop a database.
        
        Args:
            database_name: Name of the database to drop
            force: Force drop without confirmation
            
        Raises:
            psycopg2.Error: If database drop fails
        """
        # Implementation placeholder
        console.print(f"[yellow]⚠️  Would drop database '{database_name}' (force={force})[/yellow]")
        raise NotImplementedError("Database drop functionality coming soon")
    
    def dump_database(self, database_name: str, output_path: Path) -> None:
        """
        Create a dump of the specified database.
        
        Args:
            database_name: Name of the database to dump
            output_path: Path where to save the dump file
            
        Raises:
            RuntimeError: If pg_dump fails
        """
        # Implementation placeholder
        console.print(f"[yellow]⚠️  Would dump database '{database_name}' to {output_path}[/yellow]")
        raise NotImplementedError("Database dump functionality coming soon")
    
    def restore_database(self, database_name: str, dump_path: Path) -> None:
        """
        Restore a database from a dump file.
        
        Args:
            database_name: Name of the database to restore to
            dump_path: Path to the dump file
            
        Raises:
            RuntimeError: If pg_restore fails
        """
        # Implementation placeholder
        console.print(f"[yellow]⚠️  Would restore database '{database_name}' from {dump_path}[/yellow]")
        raise NotImplementedError("Database restore functionality coming soon")
    
    def get_database_info(self, database_name: str) -> Dict[str, Any]:
        """
        Get information about a database.
        
        Args:
            database_name: Name of the database
            
        Returns:
            Dictionary with database information
        """
        # Implementation placeholder
        console.print(f"[yellow]⚠️  Would get info for database '{database_name}'[/yellow]")
        return {}
    
    def close(self) -> None:
        """Close database connection."""
        if self._connection:
            self._connection.close()
            self._connection = None 